requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-model-utils==4.2.0
sqlparse==0.4.2
tzdata==2021.5
